#19 Request, Response 클래스 분리
