+++ GODA-Client/Assets/Prefabs/Manager/APIConnector.prefab
@@ -44,4 +44,4 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b7647d4cf6c1d2144bb698c80811ca6d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  endPointSO: {fileID: 11400000, guid: e7c0fd7f0023bbe4b809dea463ad615f, type: 2}
+  endpointSO: {fileID: 11400000, guid: e7c0fd7f0023bbe4b809dea463ad615f, type: 2}
