+++ GODA-Client/Assets/Scripts/API/Request/AccountRequest.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+[Serializable]
+public class AccountRequest
+{
+	public string email;
+	public string playerId;
+	public string password;
+}
+
