+++ GODA-Client/Assets/Scripts/API/Request/AccountRequest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c6f18525a723b8d41a430fc6f0f703ba
