+++ GODA-Client/Assets/Scripts/API/Request/LoginRequest.cs
@@ -1,0 +1,8 @@
+using System;
+
+[Serializable]
+public class LoginRequest
+{
+	public string playerId;
+	public string password;
+}
