+++ GODA-Client/Assets/Scripts/API/Request/LoginRequest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 4e88006aa4b5459438f27921acf968ea
