+++ GODA-Client/Assets/Scripts/API/Request/PlayerDataRequest.cs
@@ -1,0 +1,8 @@
+﻿[System.Serializable]
+public class PlayerDataRequest
+{
+	public string Name;
+	public int Level;
+	public int Exp;
+	public int MaxExp;
+}
