+++ GODA-Client/Assets/Scripts/API/Request/PlayerDataRequest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: c59b00181c7165840aacdc03c5798cea
