+++ GODA-Client/Assets/Scripts/API/Request/PlayerNameRequest.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+[Serializable]
+public class PlayerNameRequest
+{
+	public string playerName;
+}
