+++ GODA-Client/Assets/Scripts/API/Request/PlayerNameRequest.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 7f6cf8e606be4144c96a344b7f6c19dd
