+++ GODA-Client/Assets/Scripts/API/Response/ExpLevelUpResponse.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+[Serializable]
+public class ExpLevelUpResponse
+{
+	public int userId;
+	public int newLevel;
+	public int newExp;
+}
