+++ GODA-Client/Assets/Scripts/API/Response/LoginResponse.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+[Serializable]
+public class LoginResponse
+{
+	public string sessionId;
+	public bool isNewAccount;
+}
