+++ GODA-Client/Assets/Scripts/API/Response/LoginResponse.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a58d030190c9984478eed9e051be03cb
