+++ GODA-Client/Assets/Scripts/API/Response/PlayerLevelResponse.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+[Serializable]
+public class PlayerLevelResponse
+{
+	public int Id;
+	public int level;
+	public int exp;
+}
