+++ GODA-Client/Assets/Scripts/API/Response/PlayerNameResponse.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+[Serializable]
+public class PlayerNameResponse
+{
+	public int id;
+	public string playerName;
+}
