+++ GODA-Client/Assets/Scripts/API/Response/PlayerNameResponse.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 090a46ca2b5936e448f0f94012c92c1e
