+++ GODA-Client/Assets/Scripts/API/Response/ResponseGeneric.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+public class Response<T>
+{
+	[JsonProperty("errorCode")]
+	public int ErrorCode { get; set; }
+
+	[JsonProperty("message")]
+	public string Message { get; set; }
+
+	[JsonProperty("data")]
+	public T Data { get; set; }
+}
