+++ GODA-Client/Assets/Scripts/API/Response/ResponseGeneric.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b82f28b9945d49b4595dac0aa37cd4e8
