+++ GODA-Client/Assets/Scripts/API/Response/SignUpResponse.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+[Serializable]
+public class SignUpResponse
+{
+	public string message;
+	public int id;
+}
