+++ GODA-Client/Assets/Scripts/API/Response/SignUpResponse.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 75f41c70148988f4cbd289c2a93ff2f4
